creating Learn feature
